+++ lib/maze/server.rb
@@ -93,6 +93,8 @@ module Maze
           uploads
         when 'sourcemap', 'sourcemaps'
           sourcemaps
+        when 'ignored', 'ignored requests'
+          ignored_requests
         when 'invalid', 'invalid requests'
           invalid_requests
         else
@@ -169,6 +171,10 @@ module Maze
       # @return [RequestList] Commands to be performed
       def commands
         @commands ||= RequestList.new
+      end
+
+      def ignored_requests
+        @ignored_requests ||= RequestList.new
       end
 
       # Whether the server thread is running
@@ -270,6 +276,7 @@ module Maze
         sampling_requests.clear
         traces.clear
         logs.clear
+        ignored_requests.clear
         invalid_requests.clear
       end
     end