+++ lib/maze/servlets/servlet.rb
@@ -159,7 +159,25 @@ module Maze
       private
 
       def add_request(request)
-        Server.list_for(@request_type).add(request)
+        if @request_type == :errors and system_generated_anr?(request)
+          $logger.info 'Ignoring system-generated ANR'
+          Server.list_for('ignored').add(request)
+        else
+          Server.list_for(@request_type).add(request)
+        end
+      end
+
+      def system_generated_anr?(request)
+        body = request[:body]
+        error_class = Maze::Helper.read_key_path(body, 'events.0.exceptions.0.errorClass')
+        return false unless error_class == 'ANR'
+
+        stack_trace = Maze::Helper.read_key_path(body, 'events.0.exceptions.0.stacktrace')
+        return false unless stack_trace.kind_of?(Array)
+
+        method_1 = stack_trace.any? { |frame| frame['method'] == 'android.os.BinderProxy.transact' }
+        method_2 = stack_trace.any? { |frame| frame['method'] == 'android.app.IActivityManager$Stub$Proxy.handleApplicationCrash' }
+        method_1 && method_2
       end
 
       def log_request(request)